requirements.txt
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
yfinance>=0.2
requests>=2.31
PyYAML>=6.0
pydantic>=2.5
joblib>=1.3
plotly>=5.18
finnhub-python>=2.4
openai>=1.52
vaderSentiment>=3.3.2
tabulate>=0.9
